refactor(heartbeat): extract active-hours logic to dedicated module
Move the active-hours helpers (resolveActiveHoursTimezone,
parseActiveHoursTime, resolveMinutesInTimeZone, isWithinActiveHours)
into heartbeat-active-hours.ts to reduce heartbeat-runner.ts below
the 1030-line CI threshold after the model-override addition.
